Add curl change password ass auth events/api/ui
Added new files for the events and api and ui. as well as finished
adding the curl requests with change password. So far the curl all works
The ajax does not giving an error that the api.___ (anything) in not
a function even with the api file there and filled.